fix: pieDataLabelsPlugin typing
